requirement.txt
django
psycopg2
psycopg2-binary
GDAL
django-leaflet
gunicorn
django-heroku
whitenoise